requirements.txt
fastapi[all]
tinydb
pydantic
uvicorn
pyjwt
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
